feat: Stun particle

+++ common/src/main/java/com/yummy/naraka/jade/NarakaEntityDataProvider.java
@@ -39,7 +39,7 @@ public class NarakaEntityDataProvider implements IEntityComponentProvider {
             double lockedHealth = LockedHealthHelper.get(livingEntity);
             if (lockedHealth > 0) {
                 tooltip.add(JadeUI.smallItem(NarakaItems.LOCKED_HEALTH.get().getDefaultInstance()));
-                tooltip.append(Component.translatable(LanguageKey.JADE_STIGMA_KEY, lockedHealth));
+                tooltip.append(Component.translatable(LanguageKey.JADE_STIGMA_KEY, (int) lockedHealth));
             }
         }
     }
